requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
aiohttp==2.2.5
websockets==3.3
motor==1.1
beautifulsoup4==4.6.0
uvloop==0.8.0
#Earl-ETF==2.0.2
voluptuous==0.10.5
itsdangerous==0.24
pysha3==1.0.2
Pillow=4.2.1
